# Doc/openapi.yaml
openapi: 3.0.3
info:
  title: LLT Assistant API
  description: |
    **Unified API Specification for LLT Assistant VSCode Extension.**

    **Core Philosophy: Frontend-First & Leaky Abstraction Avoidance**
    - The Backend provides business capabilities, not atomic code operations.
    - The Frontend handles local context (Git diffs, XML parsing, File reading).
    - Async workflows are used for heavy LLM operations.

    **Feature Map:**
    - **Feature 1 (Test Generation):** `/workflows/generate-tests` (Async)
    - **Feature 2 (Coverage Optimization):** `/optimization/coverage` (Async)
    - **Feature 3 (Impact Analysis):** `/analysis/impact` (Sync/Fast)
    - **Feature 4 (Quality Analysis):** `/quality/analyze` (Sync/Batch)
  version: 1.1.0
  contact:
    name: LLT Assistant Technical Team

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.llt-assistant.com
    description: Production server

tags:
  - name: System
    description: Health and System status
  - name: Feature 1 - Generation
    description: Generate new tests from scratch or descriptions
  - name: Feature 2 - Coverage
    description: Analyze gaps and generate supplementary tests
  - name: Feature 3 - Impact
    description: Analyze code changes to find affected tests
  - name: Feature 4 - Quality
    description: Linting and quality analysis with auto-fixes
  - name: Tasks
    description: Polling for async workflows

paths:
  # ============================================================================
  # SYSTEM
  # ============================================================================
  /health:
    get:
      summary: System Health Check
      tags:
        - System
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy, degraded]
                  version:
                    type: string

  # ============================================================================
  # FEATURE 1: Test Generation (Preserved)
  # ============================================================================
  /workflows/generate-tests:
    post:
      summary: Generate tests (Async)
      description: |
        Triggers the test generation workflow.
        Used for:
        1. New test generation (Feature 1).
        2. Regenerating broken tests detected by Impact Analysis (Feature 3).
      tags:
        - Feature 1 - Generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateTestsRequest'
      responses:
        '202':
          description: Task accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ============================================================================
  # FEATURE 2: Coverage Optimization (Optimized)
  # ============================================================================
  /optimization/coverage:
    post:
      summary: Optimize coverage (Async)
      description: |
        Frontend parses `coverage.xml` locally and sends uncovered ranges.
        Backend generates targeted test cases to fill these gaps.
      tags:
        - Feature 2 - Coverage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoverageOptimizationRequest'
      responses:
        '202':
          description: Optimization task started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ============================================================================
  # FEATURE 3: Impact Analysis (Optimized)
  # ============================================================================
  /analysis/impact:
    post:
      summary: Analyze Test Impact
      description: |
        **Read-Only Analysis.**
        Frontend provides file changes (Git diffs) and relevant test file contents.
        Backend determines which tests are impacted and why.

        *Note: To fix/regenerate tests, use `/workflows/generate-tests`.*
      tags:
        - Feature 3 - Impact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImpactAnalysisRequest'
      responses:
        '200':
          description: Analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImpactAnalysisResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ============================================================================
  # FEATURE 4: Quality Analysis (Optimized)
  # ============================================================================
  /quality/analyze:
    post:
      summary: Batch Quality Analysis
      description: |
        Analyzes multiple files for defects and redundancies.
        Returns issues with **embedded fix suggestions** for zero-latency UI interactions.
      tags:
        - Feature 4 - Quality
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QualityAnalysisRequest'
      responses:
        '200':
          description: Analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityAnalysisResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ============================================================================
  # SHARED TASK POLLING
  # ============================================================================
  /tasks/{task_id}:
    get:
      summary: Poll Async Task Status
      description: Retrieve status and results for Feature 1 and Feature 2 workflows.
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task status or result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponse'
        '404':
          description: Task not found

# ============================================================================
# COMPONENTS
# ============================================================================
components:
  schemas:
    # ----------------------------------------
    # Shared Schemas
    # ----------------------------------------
    AsyncJobResponse:
      type: object
      required:
        - task_id
        - status
      properties:
        task_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing]
        estimated_time_seconds:
          type: integer

    TaskStatusResponse:
      type: object
      required:
        - task_id
        - status
      properties:
        task_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        created_at:
          type: string
          format: date-time
        result:
          oneOf:
            - $ref: '#/components/schemas/GenerateTestsResult'
            - $ref: '#/components/schemas/CoverageOptimizationResult'
          description: |
            Task result varies by feature type:
            - Feature 1 (Test Generation): Returns GenerateTestsResult with generated_code and explanation
            - Feature 2 (Coverage Optimization): Returns CoverageOptimizationResult with recommended_tests array
            Only present when status is 'completed'.
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string

    # ----------------------------------------
    # Feature 1 Schemas
    # ----------------------------------------
    GenerateTestsRequest:
      type: object
      required:
        - source_code
      properties:
        source_code:
          type: string
          description: The Python source code to test.
        user_description:
          type: string
          description: Optional hint or requirement from user.
        existing_test_code:
          type: string
          description: Optional existing test code (context for regeneration).
        context:
          type: object
          description: Extra context if triggered by Feature 3 (Regeneration).
          properties:
            mode:
              type: string
              enum: [new, regenerate]
              default: new
            target_function:
              type: string

    GenerateTestsResult:
      type: object
      properties:
        generated_code:
          type: string
          description: The complete generated test file content.
        explanation:
          type: string

    # ----------------------------------------
    # Feature 2 Schemas (Optimized)
    # ----------------------------------------
    CoverageOptimizationRequest:
      type: object
      required:
        - source_code
        - uncovered_ranges
      properties:
        source_code:
          type: string
          description: Target source file content.
        existing_test_code:
          type: string
          description: Current test file content.
        uncovered_ranges:
          type: array
          description: Ranges parsed by Frontend from coverage.xml.
          items:
            type: object
            properties:
              start_line:
                type: integer
              end_line:
                type: integer
              type:
                type: string
                enum: [line, branch]
        framework:
          type: string
          enum: [pytest, unittest]
          default: pytest

    CoverageOptimizationResult:
      type: object
      properties:
        recommended_tests:
          type: array
          items:
            type: object
            properties:
              test_code:
                type: string
                description: The code snippet for the new test case.
              target_line:
                type: integer
                description: Recommended line number to insert the ghost text.
              scenario_description:
                type: string
              expected_coverage_impact:
                type: string
                example: "Covers exception handling in calculate_tax"

    # ----------------------------------------
    # Feature 3 Schemas (Optimized)
    # ----------------------------------------
    ImpactAnalysisRequest:
      type: object
      required:
        - project_context
      properties:
        project_context:
          type: object
          required:
            - files_changed
          properties:
            files_changed:
              type: array
              items:
                type: object
                required:
                  - path
                  - diff_content
                properties:
                  path:
                    type: string
                    example: "src/utils.py"
                  diff_content:
                    type: string
                    description: "Raw git diff string for this file."
            related_tests:
              type: array
              description: |
                Frontend uses static analysis (regex/AST) to guess which test files might be relevant
                and sends their content here for the LLM to confirm impact.
              items:
                type: object
                required:
                  - path
                  - content
                properties:
                  path:
                    type: string
                  content:
                    type: string

    ImpactAnalysisResponse:
      type: object
      required:
        - impacted_tests
      properties:
        impacted_tests:
          type: array
          items:
            type: object
            properties:
              file_path:
                type: string
                description: Path to the test file.
              test_case_name:
                type: string
                description: Name of the specific test function affected.
              severity:
                type: string
                enum: [critical, high, medium, low]
                description: Used for UI coloring (Red/Orange/Yellow).
              reason:
                type: string
                description: Human-readable reason (e.g., "Function signature changed").
        suggested_action:
          type: string
          enum: [none, regenerate]
          description: Recommendation for the UI button state.

    # ----------------------------------------
    # Feature 4 Schemas (Optimized)
    # ----------------------------------------
    QualityAnalysisRequest:
      type: object
      required:
        - files
      properties:
        files:
          type: array
          items:
            type: object
            required:
              - path
              - content
            properties:
              path:
                type: string
              content:
                type: string
        mode:
          type: string
          enum: [fast, deep, hybrid]
          default: hybrid
          description: |
            fast: Rules only (e.g., On-Save).
            deep: LLM only.
            hybrid: Recommended default.

    QualityAnalysisResponse:
      type: object
      required:
        - issues
        - summary
      properties:
        analysis_id:
          type: string
        summary:
          type: object
          properties:
            total_files:
              type: integer
            total_issues:
              type: integer
            critical_issues:
              type: integer
        issues:
          type: array
          items:
            $ref: '#/components/schemas/QualityIssue'

    QualityIssue:
      type: object
      required:
        - file_path
        - line
        - severity
        - message
        - code
      properties:
        file_path:
          type: string
        line:
          type: integer
          description: 1-based line number.
        column:
          type: integer
          default: 0
        severity:
          type: string
          enum: [error, warning, info]
        code:
          type: string
          example: "redundant-assertion"
        message:
          type: string
        detected_by:
          type: string
          enum: [rule, llm]
        suggestion:
          $ref: '#/components/schemas/FixSuggestion'

    FixSuggestion:
      type: object
      nullable: true
      properties:
        type:
          type: string
          enum: [replace, delete, insert]
        new_text:
          type: string
          description: The code to apply.
        description:
          type: string
          description: 'Menu label (e.g., "Fix: Remove redundancy").'

  responses:
    BadRequest:
      description: Invalid request format or parameters.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: object